Move extracted helper method to parent class
