+++ packages/framework/src/Console/Commands/MakePublicationCommand.php
@@ -175,15 +175,4 @@ class MakePublicationCommand extends ValidatingCommand implements CommandHandleI
 
         return $pubType;
     }
-
-    protected function handleException(Exception $exception): int
-    {
-        if ($exception->getFile() === __FILE__) {
-            $this->error("Error: {$exception->getMessage()}");
-        } else {
-            $this->error("Error: {$exception->getMessage()} at {$exception->getFile()}:{$exception->getLine()}");
-        }
-
-        return Command::FAILURE;
-    }
 }
